separate storage v1 term processing
