+++ tkcore/term/main.cpp
@@ -15,7 +15,7 @@ int main(int argc, char **argv) {
 
 
     if (argc <= 1) {
-        cmd_pr::printHelp();
+        thekey_term::printHelp();
         return 0;
     }
 
@@ -23,18 +23,18 @@ int main(int argc, char **argv) {
     while (ch = getopt_long(argc, argv, short_options, long_options, NULL), ch != -1) {
         switch (ch) {
             case 'h':
-                cmd_pr::printHelp();
+                thekey_term::printHelp();
                 return 0;
             case 'f':
-                cmd_pr::findStorages(optarg);
+                thekey_term::findStorages(optarg);
                 return 0;
             case 'l':
-                cmd_pr::login(optarg);
+                thekey_term::login(optarg);
                 return 0;
             case ':':
                 switch (optopt) {
                     case 'f':
-                        cmd_pr::findStorages(optarg);
+                        thekey_term::findStorages(optarg);
                         return 0;
                     default:
                         std::cerr << "invalid option: " << (char) optopt << std::endl;
