+++ tkcore/term/storage_v1.h
@@ -1,0 +1,17 @@
+//.cpp
+// Created by panda on 14.01.24.
+//
+
+#ifndef THEKEY_STORAGE_V1_H
+#define THEKEY_STORAGE_V1_H
+
+
+#include "def_header.h"
+
+namespace thekey_term_v1 {
+
+    void login(const std::string &filePath);
+
+}
+
+#endif //THEKEY_STORAGE_V1_H
